requirements/local.txt
-r base.txt

django-extensions==2.2.5
factory-boy==2.12.0

ipdb==0.11
